// generated/models/WavesOverviewType.ts
/**
 * Seize API
 * This is the API interface description for the most commonly used operations in Seize API.  Some modifying endpoints require an authentication token.   We are in the process of documenting all Seize APIs.   If there is an API that you need, please ping us in Discord and we will aim to prioritize its documentation.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export enum WavesOverviewType {
    Latest = 'LATEST',
    MostSubscribed = 'MOST_SUBSCRIBED',
    HighLevelAuthor = 'HIGH_LEVEL_AUTHOR',
    AuthorYouHaveRepped = 'AUTHOR_YOU_HAVE_REPPED',
    MostDropped = 'MOST_DROPPED',
    MostDroppedByYou = 'MOST_DROPPED_BY_YOU',
    RecentlyDroppedTo = 'RECENTLY_DROPPED_TO',
    RecentlyDroppedToByYou = 'RECENTLY_DROPPED_TO_BY_YOU'
}